feat: move calculator to shared components
