--- src/components/pages/variable/example/index.js
@@ -1,3 +1,0 @@
-import Example from './example';
-
-export default Example;
