+++ src/components/shared/compute-calculator/images/left-glow-mobile.png
