+++ src/components/shared/compute-calculator/images/left-glow.png
