+++ src/components/shared/compute-calculator/images/right-glow-mobile.png
