+++ src/components/shared/compute-calculator/images/right-glow.png
