+++ src/components/shared/compute-calculator/index.js
@@ -1,0 +1,3 @@
+import ComputeCalculator from './compute-calculator';
+
+export default ComputeCalculator;
